pyproject.toml
[tool.poetry]
name = "env"
version = "0.1.0"
description = "Un e-commerce para micro-empresarios"
authors = ["c15-45-ft-python-react"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.0"
pillow = "^10.1.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
black = "^21.12b0"
isort = "^5.9.3"